web/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.4
python-multipart>=0.0.9
sqlalchemy>=2.0.0
pymysql>=1.1.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
cryptography>=42.0.0
slowapi>=0.1.9
itsdangerous>=2.2.0
python-dotenv>=1.0.0
